// manager_frontend/src/client/models/ManagerCustomerBranchItemResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ManagerCustomerBranchItemResponse = {
    id: number;
    customer_id: number;
    name?: (string | null);
    delivery_address: string;
    contact_name?: (string | null);
    contact_phone?: (string | null);
    is_default?: boolean;
    created_at?: (string | null);
    updated_at?: (string | null);
};